ai_scalper_bot/requirements.txt
pyyaml
websockets
python-dotenv
numpy
pandas
requests
plotly
pydantic
aiohttp
websocket-client
python-binance
psutil
xgboost
pyarrow
scikit-learn
cryptography>=42.0.0
